Split SvgPath
