+++ Source/Paths/SvgPath.cs
@@ -1,8 +1,4 @@
-﻿using System.Drawing;
-#if !NO_SDC
-using System.Drawing.Drawing2D;
-#endif
-using Svg.Pathing;
+﻿using Svg.Pathing;
 
 namespace Svg
 {
@@ -12,10 +8,6 @@ namespace Svg
     [SvgElement("path")]
     public partial class SvgPath : SvgMarkerElement, ISvgPathElement
     {
-#if !NO_SDC
-        private GraphicsPath _path;
-#endif
-
         /// <summary>
         /// Gets or sets a <see cref="SvgPathSegmentList"/> of path data.
         /// </summary>
@@ -43,53 +35,6 @@ namespace Svg
             get { return GetAttribute<float>("pathLength", false); }
             set { Attributes["pathLength"] = value; }
         }
-
-#if !NO_SDC
-        /// <summary>
-        /// Gets the <see cref="GraphicsPath"/> for this element.
-        /// </summary>
-        public override GraphicsPath Path(ISvgRenderer renderer)
-        {
-            if (_path == null || IsPathDirty)
-            {
-                _path = new GraphicsPath();
-
-                if (PathData != null && PathData.Count > 0 && PathData.First is SvgMoveToSegment)
-                {
-                    var start = PointF.Empty;
-                    foreach (var segment in PathData)
-                        start = segment.AddToPath(_path, start, PathData);
-
-                    if (_path.PointCount == 0)
-                    {
-                        if (PathData.Count > 0)
-                        {
-                            // special case with one move command only, see #223
-                            // make sure the case is valid, but nothing is drawn
-                            var segment = PathData.Last;
-                            _path.AddLine(segment.End, segment.End);
-                            Fill = SvgPaintServer.None;
-                            Stroke = SvgPaintServer.None;
-                        }
-                        else
-                            _path = null;
-                    }
-                    else if (renderer == null)
-                    {
-                        // Calculate boundary including stroke width.
-                        var radius = StrokeWidth * 2;
-                        var bounds = _path.GetBounds();
-                        _path.AddEllipse(bounds.Left - radius, bounds.Top - radius, 2 * radius, 2 * radius);
-                        _path.AddEllipse(bounds.Right - radius, bounds.Bottom - radius, 2 * radius, 2 * radius);
-                    }
-                }
-
-                if (renderer != null)
-                    IsPathDirty = false;
-            }
-            return _path;
-        }
-#endif
 
         public void OnPathUpdated()
         {
